refactor executor and upgrade to v0.2.0
