+++ README.md
@@ -4,10 +4,8 @@ Arbiter plugins to support more real scenarios.
 
 1. observer-plugins
 
-metric-server: Use metric-server as the source of metrics
-
-prometheus: Use promethes as the source of metrics
+default-plugins: Support to use metric-server and promethes as the source of metrics
 
 2. executor-plugins
 
-resource-tagger: Tag the resource based on the value from OBI
+default-plugins: Support to update resource based on OBI and OAP
