+++ executor-plugins/default-plugins/Dockerfile
@@ -1,4 +1,4 @@
-# Build resource-tagger plugin image
+# Build executor-default-plugins image
 
 FROM golang:1.18 as builder
 
@@ -10,22 +10,21 @@ WORKDIR /workspace
 # Copy the Go Modules manifests
 COPY go.mod go.mod
 COPY go.sum go.sum
-COPY build-resource-tagger.sh build-resource-tagger.sh
+COPY build-executor-default-plugins.sh build-executor-default-plugins.sh
 COPY staging/ staging/
 
 ## Copy the go source and vendor
 COPY cmd/server/server.go cmd/server/server.go
 COPY pkg/ pkg/
 
-RUN ./build-resource-tagger.sh ${CGO} ${ARCH} ${GOOS}
+RUN ./build-executor-default-plugins.sh ${CGO} ${ARCH} ${GOOS}
 
 ## Build binary
-#RUN GOPROXY=https://goproxy.io CGO_ENABLED=${CGO} GOOS=${GOOS} GOARCH=${ARCH} go build -a -o resource-tagger cmd/server/server.go
-#RUN GOPROXY=https://goproxy.io CGO_ENABLED=${CGO} GOOS=${GOOS} GOARCH=${ARCH} go build -a -o resource-tagger-client cmd/client/client.go
+#RUN GOPROXY=https://goproxy.io CGO_ENABLED=${CGO} GOOS=${GOOS} GOARCH=${ARCH} go build -a -o executor-default-plugins cmd/server/server.go
 
 ## Build image
 FROM alpine:3.15
 WORKDIR /
-COPY --from=builder /workspace/resource-tagger /usr/local/bin/resource-tagger
+COPY --from=builder /workspace/executor-default-plugins /usr/local/bin/executor-default-plugins
 
-CMD ["resource-tagger"]
+CMD ["executor-default-plugins"]
