+++ hack/lib/init.sh
@@ -6,17 +6,17 @@ ROOT_PATH=$(dirname "${BASH_SOURCE[0]}")/../..
 
 declare -A CODE_PATH=(
 	["observer-default-plugins"]="observer-plugins/default-plugins"
-	["executor-resource-tagger"]="executor-plugins/default-plugins"
+	["executor-default-plugins"]="executor-plugins/default-plugins"
 )
 
 declare -A CODE_MAIN_PATH=(
 	["observer-default-plugins"]="main.go"
-	["executor-resource-tagger"]="cmd/server/server.go"
+	["executor-default-plugins"]="cmd/server/server.go"
 )
 
 declare -A DOCKERFILE_PATH=(
 	["observer-default-plugins"]="observer-plugins/default-plugins/Dockerfile"
-	["executor-resource-tagger"]="executor-plugins/default-plugins/Dockerfile"
+	["executor-default-plugins"]="executor-plugins/default-plugins/Dockerfile"
 )
 
 source "${ROOT_PATH}/hack/lib/util.sh"
